obj: initial obj implementation
Again taking a page from the Linux kernel, I have added an "obj"
and an API to interact with it. Similar to a kobj, an obj serves
as an easy and standard way to make something reference counted
and named. Each obj also has a "type", which describes how to
unpack the container from the obj and what to do with the obj
when its reference count reaches zero. It also has a parent, for
when it and its containing struct are associated with something else.

Differing from kobj though, an obj also has a void* to "ops".
These "ops" will make more sense later, but they are essentially
used when a group of objs all have the same purpose and/or support the
same set of operations, but all have their own implementations.

It's kind of like the specific "methods" on a bunch of things that
all implement a common "interface"

Change-Id: I214716d2486e7d0d6853824ef13f77c5b4d21b69