obj: classes and collections
This commit adds 2 important features: classes and collections.

Collections are nothing more than a group of objs, but they handle
iteration and locking internally, so you don't have to. Abstracting
away the synchronization and enumeration of collections will allow
enhancements to accessing these groups without affecting the API and
anywhere it's used.

Classes are a bit starnger. The purpose of a class is to be able to
easily interact with a certain "type" of obj. A class has an obj type
and destructor associated with it, and a collection it maintains.
One can "instantiate" a class with a specific set of ops, and that
obj goes onto the classes collection. Classes are named as well.
Why would this be useful? Think about a certain type of sensor, like
an accelerometer. An "accelerometer" class would define a set of ops
that each accelerometer must support, and could maintain a named
collection of every accelerometer in the system, making it easy
to use any accelerometer. The person writing the driver for a specific
accelerometer would only need to implement the standard set of ops
for their particular chip, and instantiate a new accelerometer with
those ops from the standard class, and it would be visible to the
entire system.

Comments in header files have more info.

Change-Id: Ic59b00731ce65877f2ac7b8ccecb89831c89e27b